+++ python/BioSimSpace/Sandpit/Exscientia/FreeEnergy/_restraint.py
@@ -48,8 +48,13 @@ from ..Units.Temperature import kelvin as _kelvin
 
 
 class Restraint:
-    """The Restraint class which holds the restraint information for the ABFE
-    calculations. Currently only Boresch type restraint is supported.
+    """
+    The Restraint class which holds the restraint information for the ABFE
+    calculations. Currently Boresch restaraints and multiple distance restraints
+    (between pairs of atoms) are supported. For the multiple distance restraints,
+    it is assumed that all restraints but one will be released after decoupling,
+    and an analytical correction will be applied to account for releasing the last
+    restraint.
 
     Boresch restraint is a set of harmonic restraints containing one bond, two
     angle and three dihedrals, which comes from three atoms in the ligand
@@ -62,7 +67,7 @@ class Restraint:
     Angles: r2-r1-l1 (thetaA0, kthetaA), r1-l1-l2 (thetaB0, kthetaB)
     Dihedrals: r3-r2-r1-l1 (phiA0, kphiA), r2-r1-l1-l2 (phiB0, kphiB), r1-l1-l2-l3 (phiC0, kphiC)
 
-    The restraint_dict has the following compact format.
+    The Boresch restraint_dict has the following compact format.
 
     restraint_dict = {
         "anchor_points":{"r1": BioSimSpace._SireWrappers.Atom,
@@ -83,7 +88,33 @@ class Restraint:
                            "kphiA": BioSimSpace.Types.Energy / (BioSimSpace.Types.Area * BioSimSpace.Types.Area),
                            "kphiB": BioSimSpace.Types.Energy / (BioSimSpace.Types.Area * BioSimSpace.Types.Area),
                            "kphiC": BioSimSpace.Types.Energy / (BioSimSpace.Types.Area * BioSimSpace.Types.Area)}}
+
+    The multiple distance restraints are flat-bottom restraints, which are represented as a dictionaries
+    of the form:
+
+    distance_restraint_dict = {r1: BioSimSpace._SireWrappers.Atom,
+                               l1: BioSimSpace._SireWrappers.Atom,
+                               r0: BioSimSpace.Types.Length,
+                               kr: BioSimSpace.Types.Energy / BioSimSpace.Types.Area,
+                               r_fb: BioSimSpace.Types.Length}
+
+    where r1 and l1 are the atoms involved in the restraint in the receptor and the ligand respectively,
+    r0 and kr are the equilibrium distance and the force constant, and r_fb is the flat bottom radius.
+
+    The overall restraint_dict is a dictionary of the form:
+
+    restraint_dict = {"distance_restraints": [distance_restraint_dict, ...],
+                      "permanent_distance_restraint": distance_restraint_dict}
+
+    where distance_restraints is a list of distance restraints which are released after decoupling,
+    and permanent_distance_restraint is the distance restraint which is not released after decoupling.
     """
+
+    # Create a dict of supported restraints and compatible engines.
+    supported_restraints = {
+        "boresch": ["gromacs", "somd"],
+        "multiple_distance": ["gromacs", "somd"],
+    }
 
     def __init__(self, system, restraint_dict, temperature, restraint_type="Boresch"):
         """
@@ -102,7 +133,7 @@ class Restraint:
             The temperature of the system
 
         restraint_type : str
-            The type of the restraint. (`Boresch`, )
+            The type of the restraint. (`Boresch`, `multiple_distance`)
         """
         if not isinstance(temperature, _Temperature):
             raise ValueError(
@@ -208,10 +239,75 @@ class Restraint:
                             " values further from 0 or pi radians."
                         )
 
+        elif restraint_type.lower() == "multiple_distance":
+            self._restraint_type = "multiple_distance"
+
+            if not set(restraint_dict.keys()) == {
+                "distance_restraints",
+                "permanent_distance_restraint",
+            }:
+                raise ValueError(
+                    "restraint_dict must have keys 'distance_restraints' and 'permanent_distance_restraint'"
+                )
+
+            # Warn the user if there are no distance restraints (although they may be deliberately supplying
+            # only the permanent distance restraint)
+            if len(restraint_dict["distance_restraints"]) == 0:
+                _warnings.warn(
+                    "No distance restraints have been specified other than the permanent distance restraint."
+                )
+
+            # Check each distance restraint is of the correct format
+            all_restraints = restraint_dict["distance_restraints"] + [
+                restraint_dict["permanent_distance_restraint"]
+            ]
+            for single_restraint_dict in all_restraints:
+                if not set(single_restraint_dict.keys()) == {
+                    "r1",
+                    "l1",
+                    "r0",
+                    "r_fb",
+                    "kr",
+                }:
+                    raise ValueError(
+                        "distance_restraint_dict must have keys 'r1', 'l1', 'r0', 'r_fb', 'kr' "
+                        f"but has keys {list(single_restraint_dict.keys())}"
+                    )
+
+                # Check that the atoms are of type BioSimSpace._SireWrappers.Atom
+                for key in ["r1", "l1"]:
+                    if not isinstance(single_restraint_dict[key], _Atom):
+                        raise ValueError(
+                            f"distance_restraint_dict['{key}'] "
+                            "must be of type 'BioSimSpace._SireWrappers.Atom'"
+                        )
+
+                # Test that all quantities have the correct units
+                for key in ["r0", "r_fb"]:
+                    if not isinstance(single_restraint_dict[key], _Length):
+                        raise ValueError(
+                            f"distance_restraint_dict['{key}'] must be of type "
+                            "'BioSimSpace.Types.Length'"
+                        )
+                if not single_restraint_dict["kr"].dimensions() == (
+                    0,
+                    0,
+                    0,
+                    1,
+                    -1,
+                    0,
+                    -2,
+                ):
+                    raise ValueError(
+                        "distance_restraint_dict['kr'] must be of type "
+                        "'BioSimSpace.Types.Energy'/'BioSimSpace.Types.Length^2'"
+                    )
+
         else:
             raise NotImplementedError(
                 f"Restraint type {type} not implemented "
-                f"yet. Only Boresch restraint is supported."
+                f"yet. Only {Restraint.supported_restraints.keys()} "
+                "are currently supported."
             )
 
         self._restraint_dict = restraint_dict
@@ -236,33 +332,55 @@ class Restraint:
             raise TypeError(
                 "'system' must be of type 'BioSimSpace._SireWrappers.System'"
             )
-        else:
-            if self._restraint_type == "boresch":
-                # Check if the ligand atoms are decoupled.
-                # Find the decoupled molecule, assume that only one can be
-                # decoupled.
-                (decoupled_mol,) = system.getDecoupledMolecules()
-                for key in ["l1", "l2", "l3"]:
-                    atom = self._restraint_dict["anchor_points"][key]
-                    # Discussed in https://github.com/michellab/BioSimSpace/pull/337
-                    if (
-                        atom._sire_object.molecule().number()
-                        != decoupled_mol._sire_object.number()
-                    ):
-                        raise ValueError(
-                            f"The ligand atom {key} is not from decoupled moleucle."
-                        )
-                for key in ["r1", "r2", "r3"]:
-                    atom = self._restraint_dict["anchor_points"][key]
-                    if not atom in system:
-                        raise ValueError(
-                            f"The protein atom {key} is not in the system."
-                        )
 
-            # Store a copy of solvated system.
-            self._system = system.copy()
+        if self._restraint_type == "boresch":
+            # Check if the ligand atoms are decoupled.
+            # Find the decoupled molecule, assume that only one can be
+            # decoupled.
+            (decoupled_mol,) = system.getDecoupledMolecules()
+            for key in ["l1", "l2", "l3"]:
+                atom = self._restraint_dict["anchor_points"][key]
+                # Discussed in https://github.com/michellab/BioSimSpace/pull/337
+                if (
+                    atom._sire_object.molecule().number()
+                    != decoupled_mol._sire_object.number()
+                ):
+                    raise ValueError(
+                        f"The ligand atom {key} is not from decoupled molecule."
+                    )
+            for key in ["r1", "r2", "r3"]:
+                atom = self._restraint_dict["anchor_points"][key]
+                if not atom in system:
+                    raise ValueError(f"The receptor atom {key} is not in the system.")
 
-    def _gromacs_boresch(self):
+        if self._restraint_type == "multiple_distance":
+            # Check if the ligand atoms are decoupled.
+            # Find the decoupled molecule, assume that only one can be
+            # decoupled.
+            (decoupled_mol,) = system.getDecoupledMolecules()
+
+            all_restraints = self._restraint_dict["distance_restraints"] + [
+                self._restraint_dict["permanent_distance_restraint"]
+            ]
+            for single_restraint_dict in all_restraints:
+                ligand_atom = single_restraint_dict["l1"]
+                if (
+                    ligand_atom._sire_object.molecule().number()
+                    != decoupled_mol._sire_object.number()
+                ):
+                    raise ValueError(
+                        f"The ligand atom {ligand_atom} is not from decoupled molecule."
+                    )
+                receptor_atom = single_restraint_dict["r1"]
+                if not receptor_atom in system:
+                    raise ValueError(
+                        f"The protein atom {receptor_atom} is not in the system."
+                    )
+
+        # Store a copy of solvated system.
+        self._system = system.copy()
+
+    def _gromacs_boresch(self, perturbation_type=None):
         """Format the Gromacs string for boresch restraint."""
 
         # Format the atoms into index list
@@ -368,7 +486,130 @@ class Restraint:
 
         return "\n".join(output)
 
-    def _somd_boresch(self):
+    def _gromacs_multiple_distance(self, perturbation_type=None):
+        """
+        Format the Gromacs string for multiple distance restraints.
+
+        Parameters
+        ----------
+        perturbation_type : str, optional, default=None
+            The type of perturbation to applied during the current stage of the free energy
+            calculation. If the perturbation type is "release_restraint", the permanent distance
+            restraint will be written as a distance restraint (topology file directive [ distance
+            restraints ], not affected by any lambda values), while all other restraints will be
+            written as restraint potentials (topology file directive [ bonds ], affected by bonded-lambda.
+            This allows the bond restraints to be released while retaining the permanent distance restraint.
+            For all other perturbation types, all restraints will be written as restraint potential bonds.
+
+        Returns
+        -------
+        str
+            The Gromacs string for the multiple distance restraints.
+        """
+
+        def _get_distance_restraint_restraint_str(r1, l1, r0, r_fb, kr):
+            """
+            Get the text line specifying a distance restraint restraint
+            (unaffected by any lambdas).
+            """
+            # Calculate parameters.
+            ai = self._system.getIndex(r1) + 1
+            aj = self._system.getIndex(l1) + 1
+            low = r0 - r_fb
+            up1 = r0 + r_fb
+            up2 = 100  # Set this unresonably high so that we never get a linear potential.
+
+            # Format strings, remembering to convert units.
+            restr_type_str = "2"  # No averaging
+            restr_index_str = "0"  # Only one restraint
+            low_str = "{:.3f}".format(low / _nanometer)
+            up1_str = "{:.3f}".format(up1 / _nanometer)
+            up2_str = "{:.3f}".format(up2)
+            fac_str = (
+                "1.0"  # Scale the force constant (specified in the mdp file) by 1.0.
+            )
+
+            # Format entire string.
+            # ai, aj, type, index, type', low, up1, up2, fac
+            distance_restraint_restraint_parameter_string = f"  {ai:<10} {aj:<10} {restr_type_str:<10} {restr_index_str:<10} {restr_type_str:<10} {low_str:<10} {up1_str:<10} {up2_str:<10} {fac_str:<10}"
+
+            return distance_restraint_restraint_parameter_string
+
+        def _get_restraint_potential_bond_str(r1, l1, r0, r_fb, kr):
+            """
+            Get the text line specifying a restraint potential bond
+            (affected by bonded-lambda).
+            """
+            # Calculate parameters.
+            ai = self._system.getIndex(r1) + 1
+            aj = self._system.getIndex(l1) + 1
+            low = r0 - r_fb
+            up1 = r0 + r_fb
+            up2 = 100  # Set this unresonably high so that we never get a linear potential.
+
+            # Format strings, remembering to convert units.
+            restr_type_str = "10"  # Restraint potential bond
+            low_str = "{:.3f}".format(low / _nanometer)
+            up1_str = "{:.3f}".format(up1 / _nanometer)
+            up2_str = "{:.3f}".format(up2)
+            kdr_0_str = "{:.2f}".format(
+                0
+            )  # Force constant is 0 when bonded-lambda = 0.
+            kdr_str = "{:.2f}".format(kr / (_kj_per_mol / _nanometer**2))
+
+            # Format entire string.
+            # ai, aj, type, lowA, up1A, up2A, kdrA, lowB, up1B, up2B, kdrB
+            restraint_potential_bond_parameter_string = f"  {ai:<10} {aj:<10} {restr_type_str:<10} {low_str:<10} {up1_str:<10} {up2_str:<10} {kdr_0_str:<10} {low_str:<10} {up1_str:<10} {up2_str:<10} {kdr_str:<10}"
+
+            return restraint_potential_bond_parameter_string
+
+        # Write the output string.
+        output = [
+            "[ intermolecular_interactions ]",
+        ]
+
+        output.append("[ bonds ]")
+        output.append(
+            "; ai         aj         type       lowA       up1A       up2A       kdrA       lowB       up1B       up2B       kdrB"
+        )
+        for restraint in self._restraint_dict["distance_restraints"]:
+            output.append(
+                _get_restraint_potential_bond_str(
+                    restraint["r1"],
+                    restraint["l1"],
+                    restraint["r0"],
+                    restraint["r_fb"],
+                    restraint["kr"],
+                )
+            )
+        if perturbation_type != "release_restraint":
+            output.append(
+                _get_restraint_potential_bond_str(
+                    self._restraint_dict["permanent_distance_restraint"]["r1"],
+                    self._restraint_dict["permanent_distance_restraint"]["l1"],
+                    self._restraint_dict["permanent_distance_restraint"]["r0"],
+                    self._restraint_dict["permanent_distance_restraint"]["r_fb"],
+                    self._restraint_dict["permanent_distance_restraint"]["kr"],
+                )
+            )
+        else:  # Perturbation type is release_restraint - write the permanent restraint as a distance restraint
+            output.append("[ distance_restraints ]")
+            output.append(
+                "; ai         aj         type       index      type'      low        up1        up2        fac"
+            )
+            output.append(
+                _get_distance_restraint_restraint_str(
+                    self._restraint_dict["permanent_distance_restraint"]["r1"],
+                    self._restraint_dict["permanent_distance_restraint"]["l1"],
+                    self._restraint_dict["permanent_distance_restraint"]["r0"],
+                    self._restraint_dict["permanent_distance_restraint"]["r_fb"],
+                    self._restraint_dict["permanent_distance_restraint"]["kr"],
+                )
+            )
+
+        return "\n".join(output)
+
+    def _somd_boresch(self, perturbation_type=None):
         """Format the SOMD string for the Boresch restraints."""
 
         # Indices
@@ -420,7 +661,48 @@ class Restraint:
 
         return restr_string
 
-    def toString(self, engine=None):
+    def _somd_multiple_distance(self, perturbation_type=None):
+        """Format the SOMD string for the multiple distance restraints."""
+
+        def _add_restr_to_str(restr, restr_string):
+            """Apend the information for a single restraint to the string."""
+            # Indices
+            r1 = self._system.getIndex(restr["r1"])
+            l1 = self._system.getIndex(restr["l1"])
+            # Equilibrium value
+            r0 = restr["r0"] / _angstrom
+            # Force constant. Halve these as SOMD defines force constants as E = kx**2
+            kr = (restr["kr"] / (_kcal_per_mol / (_angstrom * _angstrom))) / 2
+            # Flat bottomed radius
+            r_fb = restr["r_fb"] / _angstrom
+
+            restr_string += f"({r1}, {l1}): ({r0}, {kr}, {r_fb}), "
+            return restr_string
+
+        standard_restr_string = "distance restraints dictionary = {"
+
+        for single_restraint in self._restraint_dict["distance_restraints"]:
+            standard_restr_string = _add_restr_to_str(
+                single_restraint, standard_restr_string
+            )
+
+        if perturbation_type == "restraint":
+            # In this case, we want all restraints to be switched on, even "permanent" ones
+            standard_restr_string = _add_restr_to_str(
+                self._restraint_dict["permanent_distance_restraint"],
+                standard_restr_string,
+            )
+            return standard_restr_string[:-2] + "}"
+        else:  # Other perturbation types, we want the permanent restraints to be constantly on
+            standard_restr_string = standard_restr_string[:-2] + "}\n"
+            permanent_restr_string = "permanent distance restraints dictionary = {"
+            permanent_restr_string = _add_restr_to_str(
+                self._restraint_dict["permanent_distance_restraint"],
+                permanent_restr_string,
+            )
+            return standard_restr_string + permanent_restr_string[:-2] + "}"
+
+    def toString(self, engine, perturbation_type=None):
         """
         The method for convert the restraint to a format that could be used
         by MD Engines.
@@ -429,43 +711,50 @@ class Restraint:
         ----------
 
         engine : str
-            The molecular dynamics engine used to generate the restraint.
-            Available options currently is "GROMACS" and "SOMD". If this argument
-            is omitted then BioSimSpace will choose an appropriate engine
-            for you.
+            The molecular dynamics engine for which to generate the restraint.
+            Available options are currently "GROMACS" or "SOMD" for Boresch restraints,
+            or "SOMD" only for multiple distance restraints.
+        perturbation_type : str, optional, default=None
+            The type of perturbation to applied during the current stage of the free energy
+            calculation. This is only used for multiple distance restraints, for which all
+            restraints are converted to standard distance restraints to allow them to be
+            turned on when the perturbation type is "restraint", but for which the permanent
+            distance restraint is always active if the perturbation type is "release_restraint"
+            (or any other perturbation type).
         """
-        if engine.strip().lower() == "gromacs":
-            if self._restraint_type == "boresch":
-                return self._gromacs_boresch()
-        elif engine.lower() == "somd":
-            if self._restraint_type == "boresch":
-                return self._somd_boresch()
-            else:
-                raise NotImplementedError(
-                    f"Restraint type {self._restraint_type} not implemented "
-                    f"yet. Only Boresch restraints are supported."
-                )
-        else:
+        to_str_functions = {
+            "boresch": {"gromacs": self._gromacs_boresch, "somd": self._somd_boresch},
+            "multiple_distance": {
+                "gromacs": self._gromacs_multiple_distance,
+                "somd": self._somd_multiple_distance,
+            },
+        }
+
+        engine = engine.strip().lower()
+        try:
+            str_fn = to_str_functions[self._restraint_type][engine]
+        except KeyError:
             raise NotImplementedError(
-                f"MD Engine {engine} not implemented "
-                f"yet. Only Gromacs and SOMD are supported."
+                f"Restraint type {self._restraint_type} not implemented "
+                f"yet for {engine}."
             )
+
+        return str_fn(perturbation_type)
 
     def getCorrection(self, method="analytical"):
         """
         Calculate the free energy of releasing the restraint
-        to the standard state volume.'''
+        to the standard state volume.
 
         Parameters
         ----------
-
         method : str
             The integration method to use for calculating the correction for
             releasing the restraint to the standard state concentration.
-            "numerical" or "analytical". Note that the analytical correction
-            can introduce errors when the restraints are weak, restrained
-            angles are close to 0 or pi radians, or the restrained distance
-            is close to 0.
+            "numerical" or "analytical". Note that the Boresch analytical
+            correction can introduce errors when the restraints are weak,
+            restrained angles are close to 0 or pi radians, or the restrained
+            distance is close to 0.
 
         Returns
         ----------
@@ -473,6 +762,16 @@ class Restraint:
             Free energy of releasing the restraint to the standard state volume,
             in kcal / mol.
         """
+        # Constants. Take .value() to avoid issues with ** and log of GeneralUnit
+        v0 = (
+            ((_Sire_meter3 / 1000) / _Sire_mole) / _Sire_angstrom3
+        ).value()  # standard state volume in A^3
+        R = (
+            _k_boltz.value() * _kcal_per_mol / _kelvin
+        ).value()  # molar gas constant in kcal mol-1 K-1
+
+        # Parameters
+        T = self.T / _kelvin  # Temperature in Kelvin
 
         if self._restraint_type == "boresch":
             # Constants. Take .value() to avoid issues with ** and log of GeneralUnit
@@ -484,7 +783,6 @@ class Restraint:
             ).value()  # molar gas constant in kcal mol-1 K-1
 
             # Parameters
-            T = self.T / _kelvin  # Temperature in Kelvin
             prefactor = (
                 8 * (_np.pi**2) * v0
             )  # In A^3. Divide this to account for force constants of 0 in the
@@ -664,7 +962,101 @@ class Restraint:
                     f"Correction method {method} is not supported. Please choose from 'numerical' or 'analytical'."
                 )
 
+        if self._restraint_type == "multiple_distance":
+            if method == "analytical":
+                raise NotImplementedError(
+                    "The analytical correction is not supported for multiple distance restraints."
+                )
+
+            else:
+
+                def _numerical_distance_integrand(r, r0, r_fb, kr):
+                    """
+                    Integrand for harmonic distance restraint.
+
+                    Parameters
+                    ----------
+                    r : float
+                        Distance to be integrated, in Angstrom
+                    r0 : float
+                        Equilibrium distance, in Angstrom
+                    r_fb : float
+                        Flat-bottomed radius, in Angstrom
+                    kr : float
+                        Force constant, in kcal mol-1 A-2
+
+                    Returns
+                    -------
+                    float
+                        Value of integrand
+
+                    Notes
+                    -----
+                    The domain of the integrand is [0, infinity], but this will
+                    be truncated to [0, 8 RT] for practicality.
+                    """
+                    r_eff = abs(r - r0) - r_fb
+                    if r_eff < 0:
+                        r_eff = 0
+                    return (r**2) * _np.exp(-(kr * r_eff**2) / (2 * R * T))
+
+                def _get_correction(r0, r_fb, kr):
+                    """
+                    Get the free energy of releasing the harmonic distance restraint.
+
+                    Parameters
+                    ----------
+                    r0 : float
+                        Equilibrium distance, in Angstrom
+                    r_fb : float
+                        Flat-bottomed radius, in Angstrom
+                    kr : float
+                        Force constant, in kcal mol-1 A-2
+
+                    Returns
+                    -------
+                    float
+                        Free energy of releasing the restraint
+
+                    Notes
+                    -----
+                    The domain of the integrand is [0, infinity], but this will
+                    be truncated to [0, 8 RT] for practicality.
+                    """
+                    dist_at_8RT = (
+                        4 * _np.sqrt((R * T) / kr) + r_fb
+                    )  # Dist. which gives restraint energy = 8 RT
+                    r_min = max(0, r0 - dist_at_8RT)
+                    r_max = r0 + dist_at_8RT
+                    integrand = lambda r: _numerical_distance_integrand(r, r0, r_fb, kr)
+                    z_r = _integrate.quad(integrand, r_min, r_max)[0]
+                    dg = -R * T * _np.log(v0 / (4 * _np.pi * z_r))
+
+                    # Attatch unit of kcal/mol
+                    dg *= _kcal_per_mol
+
+                    return dg
+
+                # Get the parameters from the permanent distance restraint, which is not released
+                r0 = (
+                    self._restraint_dict["permanent_distance_restraint"]["r0"]
+                    / _angstrom
+                )
+                r_fb = (
+                    self._restraint_dict["permanent_distance_restraint"]["r_fb"]
+                    / _angstrom
+                )
+                kr = self._restraint_dict["permanent_distance_restraint"]["kr"] / (
+                    _kcal_per_mol / _angstrom2
+                )
+                _warnings.warn(
+                    "The multiple distance restraint correction is assumes that only "
+                    "the 'permanent_distance_restraint' is active."
+                )
+                return _get_correction(r0, r_fb, kr)
+
     @property
     def correction(self):
         """Give the free energy of removing the restraint."""
-        return self.getCorrection()
+        method = "analytical" if self._restraint_type == "boresch" else "numerical"
+        return self.getCorrection(method=method)