+++ python/BioSimSpace/Sandpit/Exscientia/Protocol/_free_energy.py
@@ -112,6 +112,11 @@ class FreeEnergy(_Production, _FreeEnergyMixin):
             "charge_soft" : Perturb all charging soft atom LJ terms (i.e. 0.0->value).
             "restraint" : Perturb the receptor-ligand restraint strength by linearly
                         scaling the force constants (0.0->value).
+            "release_restraint" : Used with multiple distance restraints to release all
+                                  restraints other than the "permanent" one when the ligand
+                                  is fully decoupled. Note that lambda = 0.0 is the fully
+                                  released state, and lambda = 1.0 is the fully restrained
+                                  state (i.e. 0.0 -> value).
 
              Currently perturubation_type != "full" is only supported by
              BioSimSpace.Process.Somd.