+++ python/BioSimSpace/_Config/_amber.py
@@ -210,7 +210,19 @@ class Amber(_Config):
                         # "heavy" restraints using a non-interoperable name mask.
                         if type(restraint) is str:
                             if restraint == "backbone":
-                                restraint_mask = "@CA,C,O,N"
+                                # Determine wether the system contains protein, nucleic acid, or both.
+                                restraint_atom_names = []
+                                if self._system.nAminoAcids() > 0:
+                                    restraint_atom_names += ["N", "CA", "C", "O"]
+                                if self._system.nNucleotides() > 0:
+                                    restraint_atom_names += [
+                                        "P",
+                                        "C5'",
+                                        "C3'",
+                                        "O3'",
+                                        "O5'",
+                                    ]
+                                restraint_mask = "@" + ",".join(restraint_atom_names)
                             elif restraint == "heavy":
                                 restraint_mask = "!:WAT & !@H="
                             elif restraint == "all":