+++ python/BioSimSpace/_SireWrappers/_system.py
@@ -46,6 +46,7 @@ from ..Types import Length as _Length
 from .. import Units as _Units
 
 from ._sire_wrapper import SireWrapper as _SireWrapper
+from ._utils import _prot_res, _nucl_res, _ions
 
 from sire.mol import Select as _Select
 
@@ -1774,186 +1775,6 @@ class System(_SireWrapper):
         # Initialise the list of indices.
         indices = []
 
-        # A set of protein residues. Taken from MDAnalysis.
-        prot_res = {
-            # CHARMM top_all27_prot_lipid.rtf
-            "ALA",
-            "ARG",
-            "ASN",
-            "ASP",
-            "CYS",
-            "GLN",
-            "GLU",
-            "GLY",
-            "HSD",
-            "HSE",
-            "HSP",
-            "ILE",
-            "LEU",
-            "LYS",
-            "MET",
-            "PHE",
-            "PRO",
-            "SER",
-            "THR",
-            "TRP",
-            "TYR",
-            "VAL",
-            "ALAD",
-            ## 'CHO','EAM', # -- special formyl and ethanolamine termini of gramicidin
-            # PDB
-            "HIS",
-            "MSE",
-            # from Gromacs 4.5.3 oplsaa.ff/aminoacids.rtp
-            "ARGN",
-            "ASPH",
-            "CYS2",
-            "CYSH",
-            "QLN",
-            "PGLU",
-            "GLUH",
-            "HIS1",
-            "HISD",
-            "HISE",
-            "HISH",
-            "LYSH",
-            # from Gromacs 4.5.3 gromos53a6.ff/aminoacids.rtp
-            "ASN1",
-            "CYS1",
-            "HISA",
-            "HISB",
-            "HIS2",
-            # from Gromacs 4.5.3 amber03.ff/aminoacids.rtp
-            "HID",
-            "HIE",
-            "HIP",
-            "ORN",
-            "DAB",
-            "LYN",
-            "HYP",
-            "CYM",
-            "CYX",
-            "ASH",
-            "GLH",
-            "ACE",
-            "NME",
-            # from Gromacs 2016.3 amber99sb-star-ildn.ff/aminoacids.rtp
-            "NALA",
-            "NGLY",
-            "NSER",
-            "NTHR",
-            "NLEU",
-            "NILE",
-            "NVAL",
-            "NASN",
-            "NGLN",
-            "NARG",
-            "NHID",
-            "NHIE",
-            "NHIP",
-            "NTRP",
-            "NPHE",
-            "NTYR",
-            "NGLU",
-            "NASP",
-            "NLYS",
-            "NPRO",
-            "NCYS",
-            "NCYX",
-            "NMET",
-            "CALA",
-            "CGLY",
-            "CSER",
-            "CTHR",
-            "CLEU",
-            "CILE",
-            "CVAL",
-            "CASF",
-            "CASN",
-            "CGLN",
-            "CARG",
-            "CHID",
-            "CHIE",
-            "CHIP",
-            "CTRP",
-            "CPHE",
-            "CTYR",
-            "CGLU",
-            "CASP",
-            "CLYS",
-            "CPRO",
-            "CCYS",
-            "CCYX",
-            "CMET",
-            "CME",
-            "ASF",
-        }
-
-        # A list of ion elements.
-        ions = [
-            "F",
-            "Cl",
-            "Br",
-            "I",
-            "Li",
-            "Na",
-            "K",
-            "Rb",
-            "Cs",
-            "Mg",
-            "Tl",
-            "Cu",
-            "Ag",
-            "Be",
-            "Cu",
-            "Ni",
-            "Pt",
-            "Zn",
-            "Co",
-            "Pd",
-            "Ag",
-            "Cr",
-            "Fe",
-            "Mg",
-            "V",
-            "Mn",
-            "Hg",
-            "Cd",
-            "Yb",
-            "Ca",
-            "Sn",
-            "Pb",
-            "Eu",
-            "Sr",
-            "Sm",
-            "Ba",
-            "Ra",
-            "Al",
-            "Fe",
-            "Cr",
-            "In",
-            "Tl",
-            "Y",
-            "La",
-            "Ce",
-            "Pr",
-            "Nd",
-            "Sm",
-            "Eu",
-            "Gd",
-            "Tb",
-            "Dy",
-            "Er",
-            "Tm",
-            "Lu",
-            "Hf",
-            "Zr",
-            "Ce",
-            "U",
-            "Pu",
-            "Th",
-        ]
-
         # Whether we've searched a perturbable system. If so, then we need to process
         # the search results differently. (There will be one for each molecule.)
         is_perturbable_system = False
@@ -1964,11 +1785,12 @@ class System(_SireWrapper):
             if self.nPerturbableMolecules() == 0:
                 # Backbone restraints.
                 if restraint == "backbone":
-                    # Find all N, CA, C, and O atoms in protein residues.
+                    # Find all backbone atoms in protein residues or nucleotides.
                     string = (
                         "(not water) and (resname "
-                        + ",".join(prot_res)
-                        + ") and (atomname N,CA,C,O)"
+                        + ",".join(_prot_res)
+                        + ",".join(_nucl_res)
+                        + ") and (atomname N,CA,C,O,P,/C5'/,/C3'/,/O3'/,/O5'/)"
                     )
                     try:
                         search = self.search(string, property_map)
@@ -1977,7 +1799,7 @@ class System(_SireWrapper):
 
                 elif restraint == "heavy":
                     # Convert to a formatted string for the search.
-                    ion_string = ",".join(ions + ["H,Xx"])
+                    ion_string = ",".join(_ions + ["H,Xx"])
                     # Find all non-water, non-hydrogen, non-ion elements.
                     string = f"(not water) and (not element {ion_string})"
                     try:
@@ -1987,7 +1809,7 @@ class System(_SireWrapper):
 
                 elif restraint == "all":
                     # Convert to a formatted string for the search.
-                    ion_string = ",".join(ions)
+                    ion_string = ",".join(_ions)
                     # Find all non-water, non-ion elements.
                     string = f"(not water) and (not element {ion_string})"
                     try:
@@ -2018,11 +1840,12 @@ class System(_SireWrapper):
 
                     if restraint == "backbone":
                         if not mol.isWater():
-                            # Find all N, CA, C, and O atoms in protein residues.
+                            # Find all backbone atoms in protein residues or nucleotides.
                             string = (
-                                "(resname "
-                                + ",".join(prot_res)
-                                + ") and (atomname N,CA,C,O)"
+                                "(not water) and (resname "
+                                + ",".join(_prot_res)
+                                + ",".join(_nucl_res)
+                                + ") and (atomname N,CA,C,O,P,/C5'/,/C3'/,/O3'/,/O5'/)"
                             )
                             try:
                                 search = mol.search(string, _property_map)
@@ -2032,7 +1855,7 @@ class System(_SireWrapper):
                     elif restraint == "heavy":
                         if not mol.isWater():
                             # Convert to a formatted string for the search.
-                            ion_string = ",".join(ions + ["H,Xx"])
+                            ion_string = ",".join(_ions + ["H,Xx"])
                             # Find all non-water, non-hydrogen, non-ion elements.
                             string = f"not element {ion_string}"
                             try:
@@ -2043,7 +1866,7 @@ class System(_SireWrapper):
                     elif restraint == "all":
                         if not mol.isWater():
                             # Convert to a formatted string for the search.
-                            ion_string = ",".join(ions)
+                            ion_string = ",".join(_ions)
                             # Find all non-water, non-ion elements.
                             string = f"not element {ion_string}"
                             try:
@@ -2074,9 +1897,12 @@ class System(_SireWrapper):
 
             if restraint == "backbone":
                 if not mol.isWater():
-                    # Find all N, CA, C, and O atoms in protein residues.
+                    # Find all backbone atoms in protein residues or nucleotides.
                     string = (
-                        "(resname " + ",".join(prot_res) + ") and (atomname N,CA,C,O)"
+                        "(resname "
+                        + ",".join(_prot_res)
+                        + ",".join(_nucl_res)
+                        + ") and (atomname N,CA,C,O,P,/C5'/,/C3'/,/O3'/,/O5'/)"
                     )
                     try:
                         search = mol.search(string, _property_map)
@@ -2086,7 +1912,7 @@ class System(_SireWrapper):
             elif restraint == "heavy":
                 if not mol.isWater():
                     # Convert to a formatted string for the search.
-                    ion_string = ",".join(ions + ["H,Xx"])
+                    ion_string = ",".join(_ions + ["H,Xx"])
                     # Find all non-water, non-hydrogen, non-ion elements.
                     string = f"not element {ion_string}"
                     try:
@@ -2097,7 +1923,7 @@ class System(_SireWrapper):
             elif restraint == "all":
                 if not mol.isWater():
                     # Convert to a formatted string for the search.
-                    ion_string = ",".join(ions)
+                    ion_string = ",".join(_ions)
                     # Find all non-water, non-ion elements.
                     string = f"not element {ion_string}"
                     try:
@@ -2114,7 +1940,7 @@ class System(_SireWrapper):
                 if num_results == 0:
                     msg = "No atoms matched the restraint!"
                     if restraint == "backbone":
-                        msg += " Backbone restraints only apply to atoms in protein residues."
+                        msg += " Backbone restraints only apply to atoms in protein residues or nucleotides."
                     raise _IncompatibleError(msg)
 
             # Loop over the searches for each molecule.
@@ -2131,9 +1957,7 @@ class System(_SireWrapper):
             if len(search) == 0 and not allow_zero_matches:
                 msg = "No atoms matched the restraint!"
                 if restraint == "backbone":
-                    msg += (
-                        " Backbone restraints only apply to atoms in protein residues."
-                    )
+                    msg += " Backbone restraints only apply to atoms in protein residues or nucleotides."
                 raise _IncompatibleError(msg)
 
             # Now loop over all matching atoms and get their indices.
@@ -2148,6 +1972,80 @@ class System(_SireWrapper):
 
         return indices
 
+    def getAminoAcids(self, property_map={}):
+        """
+        Return a list containing all of the amino acid residues in the system.
+
+        Parameters
+        ----------
+
+        property_map : dict
+            A dictionary that maps system "properties" to their user defined
+            values. This allows the user to refer to properties with their
+            own naming scheme, e.g. { "charge" : "my-charge" }
+
+        Returns
+        -------
+
+        residues : [:class:`Residue <BioSimSpace._SireWrappers.Residue>`]
+            The list of all amino acid residues in the system.
+        """
+        search_string = "(resname " + ",".join(_prot_res) + ")"
+        try:
+            residues = list(self.search(search_string, property_map).residues())
+        except:
+            residues = []
+        return residues
+
+    def nAminoAcids(self):
+        """
+        Return the number of amino acid residues in the system.
+
+        Returns
+        -------
+
+        num_aa : int
+            The number of amino acid residues in the system.
+        """
+        return len(self.getAminoAcids())
+
+    def getNucleotides(self, property_map={}):
+        """
+        Return a list containing all of the nucleotide residues in the system.
+
+        Parameters
+        ----------
+
+        property_map : dict
+            A dictionary that maps system "properties" to their user defined
+            values. This allows the user to refer to properties with their
+            own naming scheme, e.g. { "charge" : "my-charge" }
+
+        Returns
+        -------
+
+        residues : [:class:`Residue <BioSimSpace._SireWrappers.Residue>`]
+            The list of all nucleotide residues in the system.
+        """
+        search_string = "(resname " + ",".join(_nucl_res) + ")"
+        try:
+            residues = list(self.search(search_string, property_map).residues())
+        except:
+            residues = []
+        return residues
+
+    def nNucleotides(self):
+        """
+        Return the number of nucleotide residues in the system.
+
+        Returns
+        -------
+
+        num_nuc : int
+            The number of nucleotide residues in the system.
+        """
+        return len(self.getNucleotides())
+
     def _isParameterised(self, property_map={}):
         """
         Whether the system is parameterised, i.e. can we run a simulation
@@ -2161,6 +2059,7 @@ class System(_SireWrapper):
         property_map : dict
             A dictionary that maps system "properties" to their user defined
             values. This allows the user to refer to properties with their
+            own naming scheme, e.g. { "charge" : "my-charge" }
 
         Returns
         -------